[Bugfix][optsfile] Don't hardcode compilers to /usr/bin/g{cc|++}
The build_external.sh script generated opts files that hardcoded the
compiler to /usr/bin/g{cc|++}. That is fine in our pool, but won't work
in general. This patch allows users to switch to a different compiler by
setting the CC and CXX environment variables.

In addition, the patch makes sure to expand the environment variables to
full paths to work around a cmake problem